move MapboxOnboardRouterConfig inside NavigationOptions
